Completed Infrastructure Layer

+++ src/main/java/com/mint/project/entrypoint/controller/cardstats/dto/CardStatResponse.java
@@ -18,5 +18,5 @@ public class CardStatResponse {
 
 	int limit;
 
-	Map<String, Integer> payload;
+	Map<String, Long> payload;
 }
